Fix panic when aliasing without arguments
Calling `git town alias` without any additional arguments resulted in a panic. The validation logic used to determine if the alias argument is valid looked for the first value in the args slice, however the slice is empty when no arguments are supplied. This modification ensures that the first value of the args slice can be read, otherwise will pass an empty string to the argument validation function.