Slicken driver interface
an intent guided implementation with WasActivated, newly also provides
the initialization of the driver in one go. Therefore, additional
initialization logic can be spared, which reduces overall lines of code
and complexity.

The original reason to spreadh out the initialization logic presumably
was motivated by modifying objects for testing purposes. The next
commit resolves this adopting config mocks for tests.