Add MergeNoFastForward
Add unit tests for both MergeFastForward and MergeNoFastForward.
Commands.CommitsInPerennialBranch is used because NewCommitMessages is
more convenient for testing than the test_commands's CommitsInBranch.

Commands.CommitsInBranch(branch, None) ignores the branch name as it
is just a wrapper around CommitsInPerennialBranch if the parent is
None, so it is more readable to just call the CommitsInPerennialBranch
directly.

See #4381.