Add always-merge ship strategy
Tests are a modified copy of fast-forward tests.

Currently undo only restores the feature branch, similar to fast-forward
strategy. There is no undo support for the parent branch like in
squash-merge.

Undoing merge commits generally should be avoided. Reverting a merge
creates a new revert commit that reverts the data, but the original
commits are still part of the history. As a result any subsequent merge
will skip commits that have been merged, even though their data has been
reverted, unless the user modified the old commits in some way, e.g. by
rebasing or amending them.

While this is unlikely to be a problem in modern workflows where users
often sync their branches, we should err on the side of not doing
something that can lead to confusion potentially far in the future.

See
https://github.com/git/git/blob/master/Documentation/howto/revert-a-faulty-merge.txt
for more information.

See #4381.