Add configdomain.UseMessage
UseMessage is a variant type that supports the following options:
- `EditDefaultMessage`
- `UseCustomMessage(CommitMessage)`
- `UseDefaultMessage`

As Go doesn't support algebraic enums we can instead use helper
functions `Is<Type>()`. Compared to Go style const enums we lose the
linter's exhaustive check, so we have to use a default clause instead.

One of the advantages of this approach is we don't have to think about
`useDefaultMessage` value when we explicitly set a custom commit
message, e.g. in `test/cucumber/steps.go`. That value would be unused
anyway.

Alternatives considered:
1. `configdomain.UseDefaultMessage(bool)`. This name is confusing when
   used together with a custom message. For example `Commit(message,
   UseDefaultMessage(true))` is ambiguous whether a default message or a
   custom message is used.
2. `configdomain.EditCommitMessage(bool)`. This separates default
   message and editing. But currently we only care about editing default
   messages, so the caller needs to handle more complexity.
3. `configdomain.EditCommitMessage(int)`. This allows us to declare
   custom behaviours such as `EditDefaultMessage`,
   `NoEditDefaultMessage` or even `AlwaysEditMessage`,
   `EditCustomMessage` if they become necessary. However negative
   `NoEditDefaultMessage` doesn't read well, and if we just call it
   `UseDefaultMessage` we go back to (1): `Commit(message,
   UseDefaultMessage)`: do we use the message or the default message?

Looking at the issue at hand we always care about the combination of the
message and default message editing behaviour. We only care about the
default message editing if the message itself is None. So it is much
simpler to think of these two arguments as if it is a composite type,
hence UseMessage.

See https://github.com/git-town/git-town/pull/4443#discussion_r1899236258.