Make implicit continue work (#5098)
When a Git Town command encounters a previously unfinished Git Town
command, it
asks the user what to do. The user can choose to continue running the
previous
command. If that previous command needs to use the browser, that didn't
work so
far. An example is continuing a suspended `git town propose` command.

This PR gives Git Town commands the opportunity to provide a fully
initialized
forge connector to the code that continues the suspended command.