Verify that all dialog inputs defined by E2E tests are being used up (#5167)
A pain point with E2E tests is matching what dialog inputs are defined
for against what they are actually used for. This PR introduces a small
help with this: verification that the setup command assistant used up
all defined dialog inputs.

Unused dialog inputs indicated a mismatch between the setup steps that
the E2E test expects will run, and which actually run, i.e. a bug in the
E2E test.