setup assistant: make config value infrastructure optional (#5171)
Upcoming changes will skip individual screens of the setup assistent if
the
information to be entered there can be derived.

For example, if there is only one Git remote, it makes no sense to ask
the user
what the dev remote is, or whether to sync with upstream.

This PR lays the foundation for these changes by making all values that
the user
can enter optional. No changes to the actual input yet.