Verify correct E2E steps (#5177)
An ongoing pain point when maintaining E2E tests is verifying that the
steps the test defines match the steps that actually run. Currently that
needs to happen manually by visually verifying the debug output of E2E
tests.

This PR introduces an automation for this problem. Steps defined by the
E2E test get tagged what step they are for, and the Git Town executable
verifies that dialog inputs it uses are indeed for the correct step.

This has surfaced a large number of wrong tests, which this PR also
fixes.