Verify dialog test input usage at program end (#5246)
Currently we check for unused test input at the end of the setup
assistant.
There are tests where inputs are required two times: for the setup
assistant and
then for the parent branch. This doesn't currently work. This test moves
the
unused input check to the end of the execution. This surfaces a few bugs
in the
E2E tests, which this PR also fixes.