Replace custom CliConfig with PartialConfig (#5310)
The CLI config data structure is unnecessarily custom. This PR replaces
it with
a regular `PartialConfig` structure. This makes the code simpler and
more
robust, and allows providing more configuration values via CLI args in
the
future.

This enables distinguishing whether the user has provided a CLI flag or
not.

It also enables using values for `DryRun` and `Verbose` from the regular
configuration. This in turn enables providing these values in Git
metadata, the config file, or environment variables in the future.