Remove redundant explanation text (#5369)
I'm not sure this additional explanation is really necessary, given that
the stack display is pretty self-explanatory and we add a link in the
footer to the documentation for this feature on git-town.com. When in
doubt, leave it out. We can easily add it back later when we have actual
data and user feedback to justify its existence.