Move stack lineage logic to forge package (#5384)
Currently the logic for calculating the stack lineage is in the
configdomain
package. That package contains type definitions for configuration data.
A better
package is `forge`, which contains logic around hosting sourcecode on
forges.

This PR moves the definitions of data structures around stack lineage
into the `forgedomain` package, and logic around this into the `forge`
package. Separating these two packages is necessary because of the way
Go resolves dependencies between packages.