Remove unnecessary mutability from environment lookup (#5449)
Go's built-in lookup of environment variables is designed with the use
case of
minimizing overhead for importing env vars from the OS, and then lookup
up a
value or two.

It also provides mutable access to the environment variables, thereby
introducing globally shared mutable state, which makes testing this
concurrently
impossible.

Git Town will read lots of environment variables. To make this efficient
and
easy to test, this PR introduces an immutable copy of the environment
variables
optimized for O(1) lookup. This prevents an O(N*N) loop executing at the
start
of each Git Town invocation.