fix up/down command doc example graphs (#5587)
I believe the `up` and `down` command example branch graphs are
fundamentally wrong in a very confusing way. They both show the branches
being reordered in the stack, as with a `swap`, but that does not seem
to match the description of what they do nor what they do in practice.