fix go.sum hash (#5597)
it's not clear why or how this entry is out-of-date, but we've had to
add this as a patch upstream to be able to build it

```
verifying github.com/charmbracelet/x/ansi@v0.9.2: checksum mismatch
        downloaded: h1:92AGsQmNTRMzuzHEYfCdjQeUzTrgE1vfO5/7fEVoXdY=
        go.sum:     h1:VxqhWbZIXFl3/ufdpzcwlaPZRPoaOj8r+KtyeRG3C0g=

SECURITY ERROR
This download does NOT match an earlier download recorded in go.sum.
The bits may have been replaced on the origin server, or an attacker may
have intercepted the download attempt.

For more information, see 'go help module-auth'.
```