connector.SearchProposal now returns multiple results (#5767)
This PR straightens out the logic in connectors. Until now the
`SearchProposal` method was opinionated that there must always be only
one proposal for a branch, and returned an error if there were more.
This opinionatedness existed because all users of the connector always
expect only one proposal.

But this is too specific. Such opinions don't belong into a
`SearchProposal` method. The connector should return as many proposals
as it finds, and let its caller decide what to do with them.