Move the code for the proposal lineage into its own package (#5824)
Right now this code is in the `forge` package. The forge package only
implements access to code forge APIs and UIs. Embedding lineage into
proposals is something different. This PR moves this code into its own
package for now, since it is a separate, orthogonal aspect to other
parts of the codebase.